rdb_postgres/pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
pytest = "^7.4.3"

[tool.pytest.ini_options]
markers = [
    "run",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
